feat: edit android to v3
