+++ apps/web/android/app/build.gradle
@@ -4,7 +4,7 @@ android {
     namespace "kr.co.causw"
     compileSdk rootProject.ext.compileSdkVersion
     defaultConfig {
-        applicationId "kr.co.causwv2.twa"
+        applicationId "kr.co.causwv3.twa"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
@@ -31,6 +31,8 @@ repositories {
 }
 
 dependencies {
+    implementation platform('com.google.firebase:firebase-bom:33.1.2')
+    implementation 'com.google.firebase:firebase-messaging'
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation "androidx.appcompat:appcompat:$androidxAppCompatVersion"
     implementation "androidx.coordinatorlayout:coordinatorlayout:$androidxCoordinatorLayoutVersion"
